Cleaned up to a config struct and less files
